not_found_exception

+++ app/database.py
@@ -1,5 +1,5 @@
 from models import db
-
+from exceptions import NotFoundException
 
 def get_all(model):
     return model.query.all()
@@ -13,12 +13,20 @@ def add_instance(model, **kwargs):
 
 def delete_instance(model, id):
     instance = model.query.get(id)
+
+    if instance is None:
+        raise NotFoundException
+
     db.session.delete(instance)
     commit()
 
 
 def update_instance(model, id, **kwargs):
     instance = model.query.get(id)
+
+    if instance is None:
+        raise NotFoundException
+
     for attr, new_value in kwargs.items():
         setattr(instance, attr, new_value)
     commit()
